backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "resume-tailor-backend"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "anthropic>=0.27",
    "pdfplumber>=0.11",
    "PyMuPDF>=1.24",
    "sentence-transformers>=3.0",
    "weasyprint>=62",
    "jinja2>=3.1",
    "sqlalchemy>=2.0",
    "pydantic>=2.7",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
